+++ src/main/java/za/ac/cput/impl/DepartmentRepository.java
@@ -1,4 +1,4 @@
-package za.ac.cput.repository;
+package za.ac.cput.impl;
 
 public class DepartmentRepository
 {
